Add a database arg
Django's `createcachetable` allows passing in a `database` arg.

This is called when Django creates a test database, as per:

https://github.com/django/django/blob/4.1/django/db/backends/base/creation.py#L96

This line is hit when using pytest and pytest-django, however it might
also be hit in other testing setups.

The change in this commit adds a "dummy" arg to the command that fixes
this problem.